API_3_imp/TripCraft_AI/requirements.txt

fastapi
uvicorn
pydantic
python-multipart
httpx
asyncio
python-dotenv
loguru
qdrant-client
sentence-transformers
transformers
Pillow
crawl4ai
ollama
websockets
aiofiles
python-jose
passlib
bcrypt
openai-whisper # Free speech recognition
pandas
numpy
dotenv
pydantic-settings
